Sharp does not support text overlays.
If a legacy imgix image url includes text overlay (using txtfont) serverless-sharp generates an error and renders a broken image, instead of ignoring the unsupported param, which would be the preferred outcome.

This pull request adds the ‘font’ processExpectation switch and returns the processed image - ignoring the text params - allowing the app to fail gracefully and render the image with whatever params it supports.